refactor: separated register page by functions
